move firmware into a nested directory
Not sure if this is strictly necessary but it's the only way I found to
have more crates in this repo without having the config in .cargo force
the target in every one of them. A different workaround would be to use
.cargo with a specific hosted target but that won't work for everyone
(it's also just annoying not to use the default target config).

The main benefit is that now we can have the no_std child crates
executing tests in an std context.